requirements/base.txt
asgiref==3.5.2
Django==4.0.6
django-environ==0.9.0
django-ninja==0.19.1
httpx==0.23.0
psycopg2==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
tenacity==8.0.1
gunicorn==20.1.0
uvicorn==0.18.2
APScheduler==3.9.1